Empty __init__.py file which was loading models, and that can never be done in an application's __init__
